PR target/22152 	* gcc.target/i386/pr22152.c (add3): Change "count" to unsigned int.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@163927 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.target/i386/pr22152.c
@@ -4,7 +4,7 @@
 #include <mmintrin.h>
 
 __m64
-unsigned_add3 (const __m64 * a, const __m64 * b, unsigned long count)
+unsigned_add3 (const __m64 * a, const __m64 * b, unsigned int count)
 {
   __m64 sum;
   unsigned int i;
